+++ apps/website/src/docs/components/mdx-components/codeblock/react-live-block.tsx
@@ -12,7 +12,10 @@ import CodeContainer from './code-container'
 import CopyButton from './copy-button'
 import scope from './react-live-scope'
 import { liveEditorStyle, liveErrorStyle } from './styles'
-import { useState } from 'react'
+import { useEffect, useRef, useState } from 'react'
+import { Splitter } from '@ark-ui/react'
+import Frame from 'react-frame-component'
+import { FrameProvider } from './frame-provider'
 
 const features: FeaturesOptions = {
   segments: [
@@ -39,16 +42,23 @@ const features: FeaturesOptions = {
   ],
 }
 
-const LiveCodePreview = chakra(LivePreview, {
-  baseStyle: {
-    fontFamily: 'body',
-    mt: 5,
-    p: 3,
-    borderWidth: 1,
-    borderRadius: '12px',
-    fontSize: 'md',
-  },
-})
+const LiveCodePreview = chakra(LivePreview, {})
+
+const LiveCodePreviewWrapper = (props: BoxProps) => {
+  return (
+    <Box
+      fontFamily="body"
+      mt={5}
+      borderWidth="1px"
+      borderRadius="12px"
+      fontSize="md"
+      overflow="hidden"
+      {...props}
+    >
+      {props.children}
+    </Box>
+  )
+}
 
 const EditableNotice = (props: BoxProps) => {
   return (
@@ -104,12 +114,20 @@ function ReactLiveBlock({
   rawCode,
   theme,
   height,
+  inline,
   overflow,
+  overflowY,
   center,
+  padding = 4,
   ...rest
 }) {
   const [editorCode, setEditorCode] = useState(rawCode.trim())
   const onChange = (newCode) => setEditorCode(newCode.trim())
+
+  const [frameHeight, setFrameHeight] = useState<string | undefined>()
+
+  const [frameRef, setFrameRef] = useState<HTMLIFrameElement | null>(null)
+
   const liveProviderProps = {
     code: editorCode,
     scope,
@@ -119,32 +137,88 @@ function ReactLiveBlock({
 
   let sx: object = {
     overflow,
+    overflowY,
+    padding,
   }
 
   if (center) {
     sx = {
       display: 'flex',
       justifyContent: 'center',
+      alignItems: 'center',
       ...sx,
     }
   }
 
+  const [isResizing, setResizing] = useState(false)
+
+  useEffect(() => {
+    if (frameRef && !inline) {
+      setInterval(() => {
+        const height = frameRef.contentWindow?.document.body.scrollHeight ?? 240
+        setFrameHeight(height + 'px')
+      }, 500)
+    }
+  }, [frameRef, inline])
+
+  const frame = inline ? (
+    <LiveCodePreview fontSize="md" sx={sx} minH="200px" />
+  ) : (
+    <Frame ref={(ref) => setFrameRef(ref)} width="100%" height="100%">
+      <FrameProvider>
+        <LiveCodePreview fontSize="md" sx={sx} minHeight="$100vh" />
+      </FrameProvider>
+    </Frame>
+  )
+
   return (
     <FeaturesProvider value={features}>
       <LiveProvider {...liveProviderProps}>
-        <LiveCodePreview
-          zIndex="1"
-          height={height}
-          position="relative"
-          fontSize="sm"
-          sx={sx}
-        />
+        <Box
+          as={Splitter.Root}
+          display="flex"
+          alignItems="center"
+          defaultSize={[
+            { id: 'a', size: 100, minSize: 40 },
+            { id: 'b', size: 0 },
+          ]}
+          onSizeChangeStart={() => setResizing(true)}
+          onSizeChangeEnd={() => setResizing(false)}
+        >
+          <Splitter.Panel id="a">
+            <LiveCodePreviewWrapper
+              pointerEvents={isResizing ? 'none' : undefined}
+              height={height || frameHeight}
+              boxSizing="content-box"
+            >
+              {frame}
+            </LiveCodePreviewWrapper>
+          </Splitter.Panel>
+          <Box
+            as={Splitter.ResizeTrigger}
+            id="a:b"
+            width="4px"
+            bg="muted"
+            mx="2"
+            height="80px"
+            rounded="full"
+            _hover={{
+              bg: 'primary.500',
+            }}
+            _focus={{
+              outline: 'none',
+              bg: 'primary.500',
+            }}
+          />
+          <Splitter.Panel id="b"></Splitter.Panel>
+        </Box>
         {editable && <LiveError style={liveErrorStyle} />}
         <Box position="relative" zIndex="0">
           {editable && (
             <CodeContainer
               bg={theme.plain.backgroundColor}
               sx={{
+                maxHeight: '400px',
                 textarea: {
                   _focus: {
                     outline: 'none',