+++ packages/saas-ui-hooks/src/index.ts
@@ -5,3 +5,4 @@ export {
 } from './use-local-storage'
 export { useSteps } from './use-steps'
 export { type ScriptElt, type Status, useScript } from './use-script'
+export { useScrollPosition } from './use-scroll-position'