+++ packages/saas-ui-core/CHANGELOG.md
@@ -1,5 +1,37 @@
 # @saas-ui/react
 
+## 2.2.2
+
+### Patch Changes
+
+- 491315aa: Fixed forwarded ref not being passed down in IconBadge
+
+## 2.2.1
+
+### Patch Changes
+
+- 3ddca73d: Fixed issue where NavGroup collapse would not receive defaultIsOpen prop
+- Updated dependencies [4e728c26]
+  - @saas-ui/theme@2.2.1
+
+## 2.2.0
+
+### Minor Changes
+
+- 31d05ed5: Added new Navbar component 🥳
+- ebba8404: Added new IconBadge component
+- a3180b02: Removed all Component.defaultProps definitions
+
+### Patch Changes
+
+- dddb3d1a: Fixed issue where width would not be applied to the SearchInput container element
+- 91412d77: Fixed SearchInput reset when uncontrolled
+- 91412d77: Fixed theme incompatibility with Chakra UI 2.8
+- Updated dependencies [31d05ed5]
+- Updated dependencies [ebba8404]
+- Updated dependencies [91412d77]
+  - @saas-ui/theme@2.2.0
+
 ## 2.1.3
 
 ### Patch Changes