+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,29 @@
 # saas-ui-storybook
 
+## 1.1.9
+
+### Patch Changes
+
+- @saas-ui/react@2.4.2
+
+## 1.1.8
+
+### Patch Changes
+
+- @saas-ui/react@2.4.1
+
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies [31d05ed5]
+- Updated dependencies [dddb3d1a]
+- Updated dependencies [ebba8404]
+- Updated dependencies [91412d77]
+- Updated dependencies [a3180b02]
+- Updated dependencies [91412d77]
+  - @saas-ui/react@2.4.0
+
 ## 1.1.6
 
 ### Patch Changes